Move common code to lib/ package
As Openstack expects us to put our common elements here: https://docs.stackstorm.com/reference/sharing_code_sensors_actions.html
