+++ lib/openstack_connection.py
@@ -1,7 +1,7 @@
 import openstack.connection
 from openstack import connect
 
-from openstack_wrappers.missing_mandatory_param_error import MissingMandatoryParamError
+from missing_mandatory_param_error import MissingMandatoryParamError
 
 
 class OpenstackConnection:
